deps/requirements.txt
pyyaml
odsparsator==1.12.1
